Added collision handler
We only handle the case where the collision is with a wall,
in which case we bounce off of it, reversing our drive
vector in the right dimension.

For agent-agent collisions, we just let the physics engine take
care of it. So the effect is of people shouldering their way
through a crowd, basically.