New Architecture
Basically, the Game owns a Physics and a Graphics sub-object,
encapsulating that subsystem. All game objects are kept in the
Game itself, and delegate rendering and updating to their
sprites and bodies, apart from game logic.

Gone is the state system for bots. The assumption now is that,
in a distant future, there might be one team focused on game
rules, and a different team focused on making game objects look
great. The art team should figure out when a game object should
have a new appearance, based on querying the object's (read-only)
state. If that coincides with a logic change (e.g., death), that
is treated as a happy accident. Otherwise (e.g., walking north-
east) the logic team doesn't need to know whether appearance
changes or not.

Also this gives us a chance to wrap our physics engine and
rendering engine in our own wrappers.

Also new here is a LayeredArena, where we now actually control
three canvases, layered on top of each other. So 'render' is
called only once, and, based on current state, will draw to
whatever layers are relevant.