BUG: Setting body.angle to undefined breaks physics
This was being set to the object's 'aim' angle, but I guess that
was not yet set, so it ended up as undefined. I think maybe any
attempt to set body.angle directly might be ill advised, but this
was catastrophic, since it defeated all collision detection.

Now collision detection and resolution seem to be working
normall.