Add collision handling
Used pretty much exactly the old collision handling code. One
difference is that we now endow physics bodies with references
to the game object, not just an index into a list.
We could still be in trouble if the object is destroyed before
its body, but that would be a bug anyway.

Still now centralized event manager, however.

Also, although bots now take damage and die, we do not yet
remove their bodies from the physical world. TODO.