Handle death and elimination
Actor.update now returns a list of tasks for the game object to
execute. This assures that stuff that we want to have happen
simultaneously really does. And it puts the task of removing
a game object body from the physics world into the hands of
something that knows about the physics world.