Factor out behavior fragments
So now there is an Actor.prog object, including prog.main, and
this is what is called (with gameTasks list as arg) as part of
Actor.update.

Then the various behaviors/states can be changed by assigning a
different "behavior function" to prog.main.

So taking the collection of behavior functions and "onX" event
handlers together, we have a good deal of individual robot
behavior factored out of core game logic.