BUG: Bots go blind when only 2 left
The code for optimizing the case where raycasting returns only
one candidate was only setting .sight.thing, not .sight.dist.
And it's the latter that is probed to implement the RANGE
register. Fixed now.

Added a simple gun turret for pure vision testing. Also this
code sets the number of bots to 2, again for testing.