Added energy
Also moved "hardware store" properties to the 'properties' object
that is used to configure agent creation. Same with the source
code file. So that puts those settings closer to where we would
be if we got that information from a submission over the net.

Also energy for bullets is now "stacked". Storing to 'fire'
really does just store energy in the 'fire' register. Then,
at the end of an agent's turn, a single bullet is fired with
all the energy that has been stored in 'fire' during that turn.
So agents only fire a single bullet per turn, so their
bullets don't immediately collide. So we might be able to
bring back bullet-bullet collisions...