Own bullets are immune to each other
Using Matter.Body.collisionFilter to prevent an agent's own
bullets from colliding. It seems to work.

However, allowing bullets to destroy each other does not seem
to work so well. It functions fine, but leads to standoffs in
the game as each agent spends all its time shooting down the
other's bullets. So that will probably have to be changed.

Also, I think collisions are reported asymmetrically. If A
collides with B, I don't think Matter will report both (A,B)
and (B,A). Which is good! But which is also (I think) why
bullets were piling up on walls: If A was a wall, I don't think
we were handling the removal of B properly.

Also removed energy stacking: it would prevent an agent from
firing a spread of shots in a single turn. In spite of the
different aim angles, the shots would still get stacked into a
single bullet. Preventing own-bullet collisions with collision
filtering means it's not quite so important.