Forward unto moduleness
The status display widget is now getting put in its own little
module directory.

This is part of the process of adding the ability to select
participants directly from the status display. In between
games, the bot names will be replaced by <select> controls,
so you can play with the roster lineup.