Add agent selectors to status widget
In place of names, we get <select> elements. However, as yet
this has no effect on the game. The game generates the actors,
and the widget creates displays for each of them. It should be
the other way around: on game start, we collect the settings
from the widgets and use them to construct agents.