We can now select different types of agents
The RosterManager (formerly StatusDisplay) allows us to pick
different types of agent to work with.

The game then constructs agents of the selected type, positions
them as usual and runs the game.

All agents come with the same loadout, though. That should
really be a function of agent type.

Also TODO is that we can start a game, but we can't stop it.
We still have to reload in order to start a fresh game.