Changed the way the app is controlled
There are now three buttons: Reset, Start/Restart, and
Pause/Continue. They are bound to App-level methods that mostly
just delegate to corresponding Game-level methods.

This cleans up what was getting to be a bit of a mess about
restarting a game. Reset takes you back to the app's initial
state, whereas Restart allows you to keep the same roster for
a new game.

Also pause is now a button, not a keypress. There's no reason
it can't be both, but I want to stay minimalist for as long
as possible.