Event log as render target
We update the log in the same places we call render(). Someday,
we might remove the render calls, and have the Arena and
status displays be log clients. Not today, though.

There is a significant lag at the end of the game before the
log is displayed. For some reason, it is delaying the final
render, even though that is called first. Must look into this...