Display last animation frame before trying to display log
Displaying the log takes a few seconds. And for some reason it
preempts the display of the last animation frame (so you can't
actually tell who won, even).

Wrapping a timer around the call to display the log seems to do
the trick, even with a practically non-existent timeout delay.

Note that there is no visual feedback that you are waiting for
something while the log is getting ready to display, and that
wait can seem pretty long. TODO.