Enable solo mode
For development sometimes you just want to make sure your bot
doesn't just kill itself on walls, so you want it to be able to
have the arena to itself. If the game starts with only a single
agent, then the end game condition is strengthened to be
zero bots left, not just <= 1.