More information in Wall register
Now we put a numeric code in the wall register, not just a
quasi-boolean flag. So now agents can quickly tell which wall
they hit, and rig a jump table to handle it. Before, the chain
of IF statements would penalize you for hitting the last wall
in the chain. Now the response time can be more uniform.