Improve log display
If the fight is very long, then displaying the log can take a
very long time indeed. Maybe all those calls to JSON.stringify?
Anyway, I added a checkbox so that you only get the log if you
demand it. It resets on 'Reset', but not on 'Restart'. Also it
is only checked at the end of the battle, so you can flip it
on and off as many times as you want while the game is running.
Only the final state matters.

Also changed the log display textarea to be 90% of the width,
with no line wrapping. Making it a bit more readable.

Still needs some sort of progress or busy/wait indicator to
let you know something is happening.