DIV instruction was unimplemented. Navigator better wall seeking
In my attempts to make Navigators better at finding the 'closest'
wall, I implemented algorithms using the 'DIV' instruction, only
to find that it was never implemented. NOTE: This means it is
not a registered opcode, so the parser doesn't even know there's
an error. You just get some instructions with a lot of extra
arguments, if they're followed by an unregistered opcode.
This is very error prone! Need stronger checks for program
correctness during compilation!