BUG: Using random numbers causes screen refresh craziness
This commit introduces a lot of stuff, including new operators
(SIN, COS, SUB), new registers (heading, velocity, random),
and a BUGFIX: Passing local variables in function calls.

The bugfix involved passing a variable name, and then having it
be decoded in the context of the function call, not the caller.
So we now reach back in the framestack when decoding args.

But the main thing is that we now have the ability to produce
one of the weirdest behaviors I have ever seen in a programming
language. If you use a random number to initialize an agents
velocity vector, then, on every wall collision, the screen
will visibly refresh. It looks like there's two copies of the
roster manager, for a second. This is true in Chrome and Firefox,
and it remains true if you use any of a number of random number
JavaScript libraries (taking care to do nothing to the meaning
of Math.random).

The fact that this only happens on collisions (and maybe also
on startup?) suggests to me that there is an interaction with
Matter.js. I will have to provide a compatible stub package
to test this...

But wait, IT GETS WEIRDER. Merely calling the random number
generator does nothing bad, if you throw away the resulting
values. Or, if you log those values, and then paste them back
in the program as constants. You must assign the values from
the random number generator to the velocity vector in order
to trigger this behavior. It's okay if you store them in a
whole chain of local variables first.

So it's not the values the RNG produces that are the problem:
they work fine as constants. And the screen is stable until you
hit a wall. Taking damage from bullets (also a type of collision,
and requiring changes to the status display) isn't a problem.