Screen Jitter Bug: Avoid floats?
One fix is to round the inputs to setVelocity (towards zero).
No more screen jitters, though the number of initial behaviors
of the wall bouncer seem to be sharply limited.

So it seems like the nature of the numbers returned by the
RNG has something to do with this. And maybe somehow logging
them tames them somehow? Weird...