Fixed: It was the roster manager all along
More specifically, it was the fact that the roster manager/status
display was being asked to display energy values with many
decimal places. The weird flickering was because those values
were immediately replaced by nice round integers, I guess when
the energy recharged to full.

Right now, the energy value is being rounded in the roster
manager. But really we should constrain the energy register
itself to only hold integers. Better still, the energy cost
of changing direction should be rounded right where it is
being computed.