Added GTE, LTE, and IF as an alias for IFNZ
Also changed Navigator so it would be a little more careful
in approaching its target wall. That's what led to adding
these new instructions. Also testing out what we can do with
CALL.

Note that if a hardware register like sys.x is passed to a
function, and the function loops, testing the value of 'args.1',
say, it will get the current value of that register.
So, effectively, the register is being passed by reference.
Or really *by name*, I think.
It's pretty cool, but might be hard to reason about in the long
run...