Substantial changes to VML syntax
More like standard presentations of three address code, with
assignments and lvalues to the left of an equals sign.

Also labels are now 'Main:' rather than 'LABEL Main'.

Also we compile to instruction objects that are not just lists
of tokens (though the arguments of operations still are).

Amont other things, this means that function calls no longer need
the clunky ARGS...CALL... two-step of the old language.

Also, all identifiers will be case sensitive now (which really
just means opcodes).

Of course, this means the interpreter is totally broken now...