Attempting to add some rudimentary tracing
The 'debug' instruction will turn tracing on for that agent.
A trace of program execution will be collected, and displayed
at end-of-game in a textarea off to the right.

Since the debug instruction is in the code, weird things might
happen if you have two agents with the same code. (Well, you'll
get extra traces.)

It's pretty crude, but it might be a big step towards real
debugging.