Started work on AgentFactory
"Agent Kits" will now be found in public/resources/agents. For
each agent type there will be a separate directory, containing
the parts necessary to assemble an instance of that agent type.

The AgentFactory is responsible for getting those resources on
app startup, and later on will be responsible for creating the
instances from the kits.