First stab at implementing interrupts
Implemented a Wall interrupt, which triggers when you are within
a user-specified distance of a wall.

In the process, I factored out all wall code into a WallSensor
object. This includes the ability to query distance to any wall,
as well as support for the interrupt.

Added a 'Tester' agent, to test interrupt driven wall-searching
routines. Currently, it just moves to the closest wall and then
stops.