Fixed bugs relating to 'agents' interrupts
Modified 'Tester' bot to start acting like a gun turret once
it reaches a wall. Then tried to restrict its scanning to a
180 degree interval, as an optimization, including unrolling the
aim loop. Used the agents.angle register to modify the basic
every-ten-degrees aim sweep.

This led to lots of changes to correctly support reading to
and writing from paths, including in local and global variables.