BUG: Need to check if data exists before accessing its members
Bots need to check agents.data.dist to see if anything is there.
But if nothing is there, agents.data is null, so an exception
is thrown. Fixed now.

Also tried to extend the pattern where objects have an 'error'
method that gets called instead of directly throwing an exception.
That gives us the chance to decorate the error message a little
before sending it off. E.g., adding agent name to it.