Clean up wall interrupts. Bug in 'sync' instruction.
The wall sensor dispenses with flags. For edge triggering we
save the old data values themselves. This means that we can
trigger an interrupt if we were in the red zone last turn, but
now we are even further into it.

The 'sync' instruction handler was not advancing the PC.
Fixed.