The code viewer is working
That is, we can see what line we are on as we step through the
program. We don't get to see any registers or other state.

Also there's still no display of the current clock tick. Just the
chronon. So you don't know how close you are to a chronon break.