Add 'tuple' instruction
This will eventually replace 'store2'.

Writing `A = store2 B C` just looks weird. The `=` is already
doing the storing. The `store2` "operator" is really just there
to tell us not to expect a "real" operator. As such, 'tuple'
or 'pair' or 'data' or something would make more sense. I chose
'tuple'. Also, it has variable arity (>= 2), so in the long run
it should be more flexible than store2.