Add WallCrawler example
An example of a wall-crawler that doesn't use interrupts.

It depends on being able to set the old interrupt parameter by
direct assignment to an ordinary register (sys.wall.limit).
Long term, the 'irparam' instruction should go away in favor of
a register based interface. Often the sensitivity parameter is
used when interrupts themselves are not used, so it's not really
part of that system.

Also gradually converting old agent code to use new features
like infix operators and 'tuple'.