Fixed bug in GunTurret script. Support different scripts.
There was a horrible vector math bug in GunTurret, which was why
it was so hard to get targeting to behave rationally.

In this commit, we take advantage of the fact that the script to
use is specified in the config file. So a single agent can have
multiple alternative scripts. You just select which one to use
in the config file. Handy for development, and for capturing
optimizations in tutorial bots.