BUG: Conditionals, when else-branch points to location 0
If the else-branch of a conditional happens to point to
the first instruction (at position zero), then it would look
'falsy', just like if there was no else-branch. So we'd
advance to the next instruction instead of jumping to 0.

Fix: If there is no else-branch, pass -1 to doConditional().
This is an illegal program counter, so we have to be a little
careful with it, but basically it's easy to test for and to
recognize as a special indicator. So now we can jump to
instruction zero on an else-branch.