AgentsScanner registers don't disappear
Previously, the data register would be null if nothing was in
sight. However, that means that sub-registers like data.dist
could not be safely read, in case a target had drifted out of
view during the targeting routine.

Now all register data is persistent. It does not disappear, but
it may become out of date. There is a data.fresh flag to help
determine whether this data is from a new observation or not.